bot/requirements.txt
feedparser==6.0.10
python-dotenv==1.0.0
python-telegram-bot==20.4
tweepy==4.14.0



